+++ sim/Simulator/L2_miss_count.h
@@ -3,10 +3,10 @@
 
 #define MONITORING_PERIOD 8192
 
-extern uint32 aggression;
+extern int aggression;
 
 // Feedback variables
-extern uint32 accuracy;
+extern int accuracy;
 extern bool late;
 extern bool pollute;
 extern bool coverage;
@@ -15,26 +15,26 @@ extern uint32 prefetchDistance;
 extern uint32 prefetchDegree;
 
 //mshr variables
-extern int global_mshrs;
-extern int global_used_mshrs;
+extern double global_mshrs;
+extern double global_used_mshrs;
 
 //prefetch variables
-extern int global_accesses;
-extern int global_reads;
-extern int global_writebacks;
-extern int global_misses;
-extern int global_prefetchDemandMisses;
-extern int global_evictions;
-extern int global_dirty_evictions;
+extern double global_accesses;
+extern double global_reads;
+extern double global_writebacks;
+extern double global_misses;
+extern double global_prefetchDemandMisses;
+extern double global_evictions;
+extern double global_dirty_evictions;
 
-extern int global_prefetches;
-extern int global_prefetch_misses;
+extern double global_prefetches;
+extern double global_prefetch_misses;
  
-extern int global_unused_prefetches;
-extern int global_used_prefetches;
-extern int global_unreused_prefetches;
-extern int global_reused_prefetches;
-extern int global_prefetch_use_miss;
+extern double global_unused_prefetches;
+extern double global_used_prefetches;
+extern double global_unreused_prefetches;
+extern double global_reused_prefetches;
+extern double global_prefetch_use_miss;
 // Global variable counting number of L2 misses
 extern uint32 L2_miss_count;
